refactor: NotificationService 단일 클래스로 통합
